+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Make Group cards a fixed height [#5719](https://github.com/rokwire/illinois-app/issues/5719).
 
 ## [8.0.6] - 2026-02-24
 ### Added